refactor `VpnSession` and related factories to use `Int` for session indices
- **Refactor Session Index Type**:
    - Change `sessionIndex` type from `UInt` to `Int` in `VpnSession` interface and `SessionSnapshot` data class.
    - Update `BoringTunVpnSession` and `InMemoryTunnelManager` to reflect the type change across session management and telemetry.
    - Update `VpnSessionFactory` and its implementations (`BoringTunVpnSessionFactory`, `QuicVpnSessionFactory`) to accept `Int` for `sessionIndex`.

- **Update Session Factory Logic**:
    - Add a requirement in `BoringTunVpnSessionFactory` to ensure `sessionIndex` is non-negative.
    - Explicitly cast `Int` to `UInt` when interfacing with the underlying `TunnelSession` in `BoringTunVpnSessionFactory`.

- **Clean up Tests**:
    - Update `JvmBoringTunVpnSessionFactoryTest`, `InMemoryTunnelManagerDataPlaneTest`, and `InMemoryTunnelManagerTest` to use `Int` literals for session indices.

+++ new-vpn/src/commonMain/kotlin/com/rafambn/kmpvpn/session/InMemoryTunnelManager.kt
@@ -40,7 +40,7 @@ internal class InMemoryTunnelManager(
         val desiredPeers = config.peers.associateBy { peer -> peer.publicKey }
         val desiredIndexes = desiredPeers.keys
             .toSortedSet()
-            .mapIndexed { index, key -> key to (index + 1).toUInt() }
+            .mapIndexed { index, key -> key to (index + 1) }
             .toMap()
         val previousSessions = sessionsByPeer.toMap()
         val nextSessions: LinkedHashMap<String, ManagedSession> = linkedMapOf()
@@ -89,7 +89,7 @@ internal class InMemoryTunnelManager(
 
     override fun sessions(): List<SessionSnapshot> {
         return sessionsByPeer.values
-            .sortedBy { managed -> managed.session.sessionIndex.toInt() }
+            .sortedBy { managed -> managed.session.sessionIndex }
             .map { managed ->
                 SessionSnapshot(
                     peerPublicKey = managed.peer.publicKey,
@@ -104,7 +104,7 @@ internal class InMemoryTunnelManager(
 
     override fun managedSessions(): List<ManagedSession> {
         return sessionsByPeer.values
-            .sortedBy { managed -> managed.session.sessionIndex.toInt() }
+            .sortedBy { managed -> managed.session.sessionIndex }
             .map { managed -> managed.copy() }
     }
 
@@ -183,7 +183,7 @@ internal class InMemoryTunnelManager(
 
     private fun currentPeerStats(): List<VpnPeerStats> {
         return sessionsByPeer.values
-            .sortedBy { managed -> managed.session.sessionIndex.toInt() }
+            .sortedBy { managed -> managed.session.sessionIndex }
             .map { managed ->
                 val stats = peerStatsByPublicKey[managed.peer.publicKey] ?: MutablePeerStats()
                 VpnPeerStats(
@@ -214,7 +214,7 @@ internal class InMemoryTunnelManager(
     private fun shouldReuse(
         previous: ManagedSession?,
         desiredPeer: VpnPeer,
-        desiredIndex: UInt,
+        desiredIndex: Int,
     ): Boolean {
         return previous != null &&
             previous.peer == desiredPeer &&
@@ -226,7 +226,7 @@ internal class InMemoryTunnelManager(
         config: VpnConfiguration,
         peerPublicKey: String,
         desiredPeer: VpnPeer,
-        desiredIndex: UInt,
+        desiredIndex: Int,
     ): ManagedSession {
         val createdSession = sessionFactory.create(
             config = config,
